requirements.txt
# Web Scraping
crawl4ai[playwright]==0.8.0

# AI Parsing (OpenAI GPT-4o Mini)
openai==1.65.4

# Database
supabase==2.28.0

# HTTP (Telegram alerts + marketfiyati API)
requests==2.32.5

# API server
fastapi==0.128.7
uvicorn[standard]==0.40.0

# Config / Utilities
python-dotenv==1.0.1
pydantic==2.12.5
